requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pillow==11.0.0
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
setuptools==65.5.1
sniffio==1.3.1
sqlparse==0.5.2
urllib3==2.2.3
wheel==0.38.4
